added check for input file before deleting it

+++ src/TailwindPhp.php
@@ -21,7 +21,7 @@ class TailwindPhp
 
         $binFolder = dirname(__DIR__) . '/bin/';
 
-        if(!$configFile){
+        if (!$configFile) {
             $configFile = 'tailwind.config.js';
         }
 
@@ -33,7 +33,7 @@ class TailwindPhp
                 $css          = self::protectTailwindFunctionForScss($css);
                 $css          = $scssCompiler->compileString($css)->getCss();
                 $css          = self::unprotectTailwindFunctionForScss($css);
-            }catch (Throwable $e) {
+            } catch (Throwable $e) {
                 return self::error($e->getMessage(), 'To resolve the issue, look for a SCSS syntax error.');
             }
         }
@@ -52,7 +52,6 @@ class TailwindPhp
                 '-i',
                 $input,
             ], '..');
-
         } else {
 
             $tailwindcss = new Process([
@@ -61,13 +60,12 @@ class TailwindPhp
                 '-c',
                 $configFile,
             ], '..');
-
         }
 
         $status = $tailwindcss->run();
 
 
-        if($status !== 0){
+        if ($status !== 0) {
 
             $output = str_replace("\n", "\\A", $tailwindcss->getErrorOutput());
             $errors = trim(htmlspecialchars($output, ENT_COMPAT, 'UTF-8'));
@@ -77,16 +75,18 @@ class TailwindPhp
             }
 
             return self::error($errors, $solution ?? null);
-
         }
 
         // Delete tmpfile
-        unlink($input);
+        if (!empty($input)) {
+            unlink($input);
+        }
 
         return $tailwindcss->getOutput();
     }
 
-    private static function error($errors, $solution = null){
+    private static function error($errors, $solution = null)
+    {
 
         // Detect name (easter eggs)
         //$pattern = '/\/srv\/www\/vhosts\/([^\/]+)\/[^\/]+\/vendor\//';
@@ -95,7 +95,7 @@ class TailwindPhp
         $domain = (isset($_SERVER['HTTPS']) ? "https" : "http") . "://" . $_SERVER['HTTP_HOST'];
         if (preg_match($pattern, $domain, $matches)) {
             $name = ucfirst(strtolower($matches[1]));
-            $solution = "Hi $name 🌈! ". $solution;
+            $solution = "Hi $name 🌈! " . $solution;
         }
 
         $message = implode('\A', [
@@ -145,5 +145,4 @@ class TailwindPhp
     {
         return preg_replace('/\/\*TAILWINDPHP(.*?)\*\//s', '$1', $css);
     }
-
 }
